# Alter logic in Time.leap_year? for performance