Let the && short-circuit the evaluation for all years that are *not*
divisible by 4 (i.e., 75% of them!).

Can close #5572