@asterite 2018-01-11
Benchmark:

```crystal
require "benchmark"

def leap_year_old?(year)
  unless 1 <= year <= 9999
    raise ArgumentError.new "Invalid year"
  end

  (year % 4 == 0 && year % 100 != 0) || (year % 400 == 0)
end

def leap_year_new?(year)
  unless 1 <= year <= 9999
    raise ArgumentError.new "Invalid year"
  end

  year % 4 == 0 && (year % 100 != 0 || year % 400 == 0)
end

numbers = Array.new(100_000) { rand(1..9999) }

a = 0
b = 0

Benchmark.ips do |x|
  x.report("old") do
    a = numbers.count { |i| leap_year_old?(i) }
  end
  x.report("new") do
    b = numbers.count { |i| leap_year_new?(i) }
  end
end

puts a, b
```

Output:

```
old    3.6k (277.51µs) (± 5.24%)  0 B/op        fastest
new   2.66k (376.01µs) (± 3.68%)  0 B/op   1.35× slower
```

It seems slower. So 👎  for me.