@asterite 2018-01-11
Actually... when I try with all numbers from 1 to 9999 without randomness:

```
numbers = Array.new(9999) { |i| i + 1 }
```

I get:

```
old  36.49k (  27.4µs) (± 3.02%)  0 B/op   1.99× slower
new  72.79k ( 13.74µs) (± 3.15%)  0 B/op        fastest
```

Then when I try it with more numbers (uniformly distributed):

```
numbers = Array.new(100_000) { |i| (i % 9999) + 1 }
```

I get:

```
old   3.68k (271.93µs) (± 2.53%)  0 B/op   1.95× slower
new   7.18k (139.29µs) (± 4.10%)  0 B/op        fastest
```

So it seems to be faster if the numbers are uniformly distributed. But if they are random (I would suppose that's uniform too?) it's slower. I don't know what's going on.