@rab 2018-01-11
Well, If you "randomly" got a bunch of multiples of 4, then it would tend to even out.

Here's a take:
```
require "benchmark"

# n = 10_000_000_000
# Benchmark.bm do |x|
#   x.report { n.times { Time.leap_year?(rand(1..2400)) } }
# end

def leap_year_old?(year)
  unless 1 <= year <= 9999
    raise ArgumentError.new "Invalid year"
  end

  (year % 4 == 0 && year % 100 != 0) || (year % 400 == 0)
end

def leap_year_new?(year)
  unless 1 <= year <= 9999
    raise ArgumentError.new "Invalid year"
  end

  year % 4 == 0 && (year % 100 != 0 || year % 400 == 0)
end

# valid years repeated so all arrays are the same length
all_valid_years = Array.new(1_000_000) {|i| 1 + (i%9998) }
random_numbers  = Array.new(1_000_000) { rand(1..9999) }
typical_numbers = Array.new(1_000_000) { 1900 + rand(401) }

a = {} of Symbol => Int32
b = {} of Symbol => Int32

Benchmark.ips do |x|
  x.report("old valid") do
    a[:valid] = all_valid_years.count { |i| leap_year_old?(i) }
  end
  x.report("new valid") do
    b[:valid] = all_valid_years.count { |i| leap_year_new?(i) }
  end
  x.report("old random") do
    a[:random] = random_numbers.count { |i| leap_year_old?(i) }
  end
  x.report("new random") do
    b[:random] = random_numbers.count { |i| leap_year_new?(i) }
  end
  x.report("old typical") do
    a[:typical] = typical_numbers.count { |i| leap_year_old?(i) }
  end
  x.report("new typical") do
    b[:typical] = typical_numbers.count { |i| leap_year_new?(i) }
  end
end

puts a, b
```

```
$ /usr/local/bin/crystal spec/std/time/time_benchmark.cr --release
```
```
  old valid 341.52  (  2.93ms) (± 8.10%)  2.15× slower
  new valid 735.86  (  1.36ms) (± 9.61%)       fastest
 old random 369.68  (  2.71ms) (± 5.04%)  1.99× slower
 new random  271.1  (  3.69ms) (± 3.49%)  2.71× slower
old typical 353.97  (  2.83ms) (± 9.15%)  2.08× slower
new typical 271.25  (  3.69ms) (± 4.72%)  2.71× slower
{:valid => 242448, :random => 242901, :typical => 242677}
{:valid => 242448, :random => 242901, :typical => 242677}
```