@rab 2018-01-11
Just to beat this dead 🐴  a bit more. (Noting that the best, bad, and worst case numbers are based on the order of operations in `leap_year_new?`)
```
# valid years repeated so all arrays are the same length
all_valid_years = Array.new(1_000_000) { |i| 1 + (i % 9999) }
random_numbers = Array.new(1_000_000) { rand(1..9999) }
typical_numbers = Array.new(1_000_000) { 1900 + rand(401) }
best_case = Array.new(1_000_000) { 2018 }  # Not divisible by 4
bad_case = Array.new(1_000_000) { 2000 }   # divisible by 400
worst_case = Array.new(1_000_000) { 2100 } # Not divisible by 400
```
Gives: (after reordering from "best" to "worst")
```
   new best   1.07k (930.46µs) (±12.05%)       fastest
  new valid 677.23  (  1.48ms) (± 7.60%)  1.59× slower
    new bad  354.0  (  2.82ms) (± 7.65%)  3.04× slower
  new worst  353.6  (  2.83ms) (± 6.76%)  3.04× slower
old typical  350.3  (  2.85ms) (± 6.73%)  3.07× slower
 old random 348.99  (  2.87ms) (± 7.70%)  3.08× slower
    old bad 342.91  (  2.92ms) (± 6.31%)  3.13× slower
  old worst 339.02  (  2.95ms) (± 5.63%)  3.17× slower
  old valid 334.82  (  2.99ms) (± 5.28%)  3.21× slower
   old best 306.54  (  3.26ms) (±14.62%)  3.51× slower
new typical 256.27  (   3.9ms) (± 3.52%)  4.19× slower
 new random 251.54  (  3.98ms) (±11.37%)  4.27× slower
{:valid => 242424, :random => 242412, :typical => 242375, :best => 0, :bad => 1000000, :worst => 0}
{:valid => 242424, :random => 242412, :typical => 242375, :best => 0, :bad => 1000000, :worst => 0}
```